+++ api/inference/internal/ctrl/chatbot.go
@@ -4,9 +4,7 @@ import (
 	"bufio"
 	"bytes"
 	"context"
-	"crypto/sha256"
 	"crypto/tls"
-	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -20,41 +18,18 @@ import (
 	"github.com/google/uuid"
 
 	"github.com/andybalholm/brotli"
-	"github.com/ethereum/go-ethereum/accounts"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/gin-gonic/gin"
 
 	"github.com/0glabs/0g-serving-broker/common/errors"
-	teeutil "github.com/0glabs/0g-serving-broker/common/tee"
 	"github.com/0glabs/0g-serving-broker/common/util"
 	"github.com/0glabs/0g-serving-broker/inference/model"
 	"github.com/0glabs/0g-serving-broker/inference/monitor"
 )
 
-const ChatPrefix = "chat"
-
-type SigningAlgo int
-
-const (
-	ECDSA SigningAlgo = iota
-)
-
-func (r SigningAlgo) String() string {
-	return [...]string{"ecdsa"}[r]
-}
-
-type ChatSignature struct {
-	Text                string         `json:"text"`
-	SignatureEcdsa      string         `json:"signature"`
-	SigningAddressEcdsa common.Address `json:"signing_address"`
-	SigningAlgo         string         `json:"signing_algo"`
-	// Centralized provider routing proof fields (omitted for decentralized providers)
-	ProviderType       string `json:"provider_type,omitempty"`
-	ProviderIdentity   string `json:"provider_identity,omitempty"`
-	TLSCertFingerprint string `json:"tls_cert_fingerprint,omitempty"`
-}
+// ChatSignature, SigningAlgo, ChatPrefix, and the chat-signing helpers
+// (signChatWithKey, signImageResponse, signCentralizedRoutingProof, chatCacheKey)
+// live in signing.go — they are shared TEE-signing infrastructure, not
+// chatbot-specific logic.
 
 type RequestBody struct {
 	Messages []Message `json:"messages"`
@@ -330,135 +305,6 @@ func (c *Ctrl) decodeAndProcess(ctx context.Context, data []byte, encodingType s
 	}
 
 	return nil
-}
-
-func (c *Ctrl) signChatWithKey(reqBody, respData []byte, chatKey string) error {
-	hashAndEncode := func(b []byte) string {
-		h := sha256.Sum256(b)
-		return hex.EncodeToString(h[:])
-	}
-
-	requestSha256 := hashAndEncode(reqBody)
-	responseSha256 := hashAndEncode(respData)
-
-	c.logger.Debugf("requestSha256: %s, responseSha256: %s, signer address %s", requestSha256, responseSha256, c.teeService.Address.Hex())
-	text := fmt.Sprintf("%s:%s", requestSha256, responseSha256)
-	sig, err := crypto.Sign(accounts.TextHash([]byte(text)), c.teeService.ProviderSigner)
-	if err != nil {
-		return err
-	}
-
-	if sig[64] == 0 || sig[64] == 1 {
-		sig[64] += 27
-	}
-
-	chatSignature := ChatSignature{
-		Text:                text,
-		SignatureEcdsa:      hexutil.Encode(sig),
-		SigningAddressEcdsa: c.teeService.Address,
-		SigningAlgo:         ECDSA.String(),
-	}
-
-	key := c.chatCacheKey(chatKey)
-	c.logger.Debugf("key: %v, chat signature: %v", key, chatSignature)
-	c.svcCache.Set(key, chatSignature, c.chatCacheExpiration)
-	return nil
-}
-
-// signImageResponse creates a TEE signature for image generation / editing responses.
-// The signed text is: sha256(originalClientReqBody):sha256(img0),sha256(img1),...
-// This binds the signature to actual image bytes rather than provider response JSON,
-// which may contain inaccessible LAN URLs.
-func (c *Ctrl) signImageResponse(reqBody []byte, images [][]byte, chatKey string) error {
-	hashAndEncode := func(b []byte) string {
-		h := sha256.Sum256(b)
-		return hex.EncodeToString(h[:])
-	}
-
-	imgHashes := make([]string, len(images))
-	for i, img := range images {
-		imgHashes[i] = hashAndEncode(img)
-	}
-	text := hashAndEncode(reqBody) + ":" + strings.Join(imgHashes, ",")
-
-	sig, err := crypto.Sign(accounts.TextHash([]byte(text)), c.teeService.ProviderSigner)
-	if err != nil {
-		return err
-	}
-	if sig[64] == 0 || sig[64] == 1 {
-		sig[64] += 27
-	}
-
-	chatSignature := ChatSignature{
-		Text:                text,
-		SignatureEcdsa:      hexutil.Encode(sig),
-		SigningAddressEcdsa: c.teeService.Address,
-		SigningAlgo:         ECDSA.String(),
-	}
-
-	key := c.chatCacheKey(chatKey)
-	c.logger.Debugf("image signature key: %v, sig: %v", key, chatSignature)
-	c.svcCache.Set(key, chatSignature, c.chatCacheExpiration)
-	return nil
-}
-
-// signCentralizedRoutingProof creates a TEE-signed routing proof for centralized
-// provider requests. The proof includes request/response hashes, provider identity,
-// and the TLS certificate fingerprint proving the connection target.
-func (c *Ctrl) signCentralizedRoutingProof(reqBody, respData []byte, chatKey string, tlsState *tls.ConnectionState) error {
-	hashAndEncode := func(b []byte) string {
-		h := sha256.Sum256(b)
-		return hex.EncodeToString(h[:])
-	}
-
-	requestSha256 := hashAndEncode(reqBody)
-	responseSha256 := hashAndEncode(respData)
-
-	// Extract TLS certificate fingerprint — refuse to sign without it.
-	// A routing proof with an empty fingerprint carries a TEE signature but
-	// provides no TLS evidence, giving verifiers a false sense of security.
-	certInfo := teeutil.ExtractTLSInfo(tlsState)
-	if certInfo == nil || certInfo.PeerCertFingerprint == "" {
-		return fmt.Errorf("TLS certificate not available for centralized provider routing proof (response and billing are unaffected)")
-	}
-	tlsFingerprint := certInfo.PeerCertFingerprint
-	c.logger.Debugf("TLS cert fingerprint captured: %s (server: %s)", tlsFingerprint, certInfo.ServerName)
-
-	text := teeutil.FormatRoutingProofText(
-		requestSha256, responseSha256,
-		c.Service.ProviderType, c.Service.ProviderIdentity,
-		tlsFingerprint,
-	)
-
-	c.logger.Debugf("Routing proof text: %s, signer address: %s", text, c.teeService.Address.Hex())
-
-	sig, err := crypto.Sign(accounts.TextHash([]byte(text)), c.teeService.ProviderSigner)
-	if err != nil {
-		return fmt.Errorf("failed to sign routing proof: %w", err)
-	}
-
-	if sig[64] == 0 || sig[64] == 1 {
-		sig[64] += 27
-	}
-
-	chatSignature := ChatSignature{
-		Text:                text,
-		SignatureEcdsa:      hexutil.Encode(sig),
-		SigningAddressEcdsa: c.teeService.Address,
-		SigningAlgo:         ECDSA.String(),
-		ProviderType:        c.Service.ProviderType,
-		ProviderIdentity:    c.Service.ProviderIdentity,
-		TLSCertFingerprint:  tlsFingerprint,
-	}
-
-	key := c.chatCacheKey(chatKey)
-	c.logger.Debugf("key: %v, centralized chat signature: %v", key, chatSignature)
-	c.svcCache.Set(key, chatSignature, c.chatCacheExpiration)
-	return nil
-}
-
-func (*Ctrl) chatCacheKey(chatID string) string {
-	return fmt.Sprintf("%s:%s", ChatPrefix, chatID)
 }
 
 func (c *Ctrl) processSingleResponse(ctx context.Context, decodedBody []byte, outputPrice string, output *string, requestHash string, usage **Usage, isWhitelisted bool) error {
